backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
djangorestframework-simplejwt==5.3.0